Save Path Downloads

+++ include/scrapyfurto/scrapyfurto/spiders/Furtodeveiculo.py
@@ -38,39 +38,42 @@ class FurtodeveiculoSpider(scrapy.Spider):
         # Loop pelos anos e meses
         for ano in anos:
             for mes in meses:
-                # Selecionar o ano
-                ano_button = self.driver.find_element(By.ID, f'cphBody_lkAno{ano[2:]}')
-                ano_button.click()
-                time.sleep(1)  # Esperar um pouco para a próxima ação
+                if f"dados_{ano}_{mes}.csv":
+                    pass
+                else:
+                    # Selecionar o ano
+                    ano_button = self.driver.find_element(By.ID, f'cphBody_lkAno{ano[2:]}')
+                    ano_button.click()
+                    time.sleep(1)  # Esperar um pouco para a próxima ação
 
-                # Selecionar o mês
-                mes_button = self.driver.find_element(By.ID, f'cphBody_lkMes{mes}')
-                mes_button.click()
-                time.sleep(1)  # Esperar um pouco para a próxima ação
+                    # Selecionar o mês
+                    mes_button = self.driver.find_element(By.ID, f'cphBody_lkMes{mes}')
+                    mes_button.click()
+                    time.sleep(1)  # Esperar um pouco para a próxima ação
 
-                # Clicar no botão de exportar
-                export_button = self.driver.find_element(By.ID, 'cphBody_ExportarBOLink')
-                export_button.click()
-                time.sleep(10)  # Esperar o download ser concluído (ajuste conforme necessário)
+                    # Clicar no botão de exportar
+                    export_button = self.driver.find_element(By.ID, 'cphBody_ExportarBOLink')
+                    export_button.click()
+                    time.sleep(10)  # Esperar o download ser concluído (ajuste conforme necessário)
 
-                # Esperar o download ser concluído e renomear o arquivo XLS baixado
-                xls_filename_path = (f'{DOWNLOADS}/DadosBO_{ano}_{mes}(FURTO DE VEÍCULOS).xls')
-                xls_filename = pd.read_csv(xls_filename_path, encoding='UTF-16-le', sep='\t', skipinitialspace=bool, dtype=str)
-                csv_filename = f"dados_{ano}_{mes}.csv"
+                    # Esperar o download ser concluído e renomear o arquivo XLS baixado
+                    xls_filename_path = (f'{DOWNLOADS}/DadosBO_{ano}_{mes}(FURTO DE VEÍCULOS).xls')
+                    xls_filename = pd.read_csv(xls_filename_path, encoding='UTF-16-le', sep='\t', skipinitialspace=bool, dtype=str)
+                    csv_filename = f"dados_{ano}_{mes}.csv"
 
-                # Aguarde o arquivo ser baixado completamente (ajuste conforme necessário)
+                    # Aguarde o arquivo ser baixado completamente (ajuste conforme necessário)
 
-                # # Renomear o arquivo XLS baixado para incluir ano e mês no nome
-                # new_xls_filename = f"dados_{ano}_{mes}.xls"
-                # os.rename(xls_filename, new_xls_filename)
+                    # # Renomear o arquivo XLS baixado para incluir ano e mês no nome
+                    # new_xls_filename = f"dados_{ano}_{mes}.xls"
+                    # os.rename(xls_filename, new_xls_filename)
 
-                # Converter XLS para CSV usando pandas
-                xls_filename.to_csv(csv_filename, index=False)
-                # csv_data = pd.read_csv(csv_filename, sep='/t')#, engine="xlrd"
-                # csv_data.to_csv(csv_filename, index=False)
+                    # Converter XLS para CSV usando pandas
+                    xls_filename.to_csv(csv_filename, index=False)
+                    # csv_data = pd.read_csv(csv_filename, sep='/t')#, engine="xlrd"
+                    # csv_data.to_csv(csv_filename, index=False)
 
-                # Limpar o arquivo XLS após a conversão
-                os.remove(xls_filename_path)
+                    # Limpar o arquivo XLS após a conversão
+                    os.remove(xls_filename_path)
         tables = []
         for ano in anos:
             for mes in meses:
@@ -83,7 +86,7 @@ class FurtodeveiculoSpider(scrapy.Spider):
                         csv_filename, low_memory=False, index_col=None, header=0
                     )
                     tables.append(csv)
-        result = "Todas Tabelas.csv"
+        result = (f"{DOWNLOADS}/Todas Tabelas.csv")
         merge = pd.concat(tables)
         merge.to_csv(result, index=False)
